Convert koin-ktor-core into multiplatform
