+++ projects/core/koin-core/src/jvmMain/kotlin/org/koin/core/context/GlobalContext.kt
@@ -37,7 +37,7 @@ object GlobalContext : KoinContext {
 
     override fun getOrNull(): Koin? = _koin
 
-    fun getKoinApplicationOrNull(): KoinApplication? = _koinApplication
+    override fun getKoinApplicationOrNull(): KoinApplication? = _koinApplication
 
     private fun register(koinApplication: KoinApplication) {
         if (_koin != null) {
