+++ projects/ktor/koin-ktor/src/commonMain/kotlin/org/koin/ktor/ext/ApplicationExt.kt
@@ -17,13 +17,11 @@ package org.koin.ktor.ext
 
 import io.ktor.server.application.*
 import org.koin.core.Koin
-import org.koin.core.context.GlobalContext
 import org.koin.core.parameter.ParametersDefinition
 import org.koin.core.qualifier.Qualifier
-import org.koin.dsl.KoinAppDeclaration
 import org.koin.ktor.plugin.KOIN_ATTRIBUTE_KEY
-import org.koin.ktor.plugin.Koin
 import org.koin.ktor.plugin.setKoinApplication
+import org.koin.mp.KoinPlatformTools
 
 /**
  * Ktor Koin extensions
@@ -39,7 +37,8 @@ import org.koin.ktor.plugin.setKoinApplication
  */
 fun Application.getKoin(): Koin =
     attributes.getOrNull(KOIN_ATTRIBUTE_KEY)?.koin ?: run {
-        val defaultInstance = GlobalContext.getKoinApplicationOrNull() ?: error("No Koin instance started. Use install(Koin) or startKoin()")
+        val defaultContext = KoinPlatformTools.defaultContext()
+        val defaultInstance = defaultContext.getKoinApplicationOrNull() ?: error("No Koin instance started. Use install(Koin) or startKoin()")
         setKoinApplication(defaultInstance)
         attributes[KOIN_ATTRIBUTE_KEY].koin
     }
